% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tryout.R
\docType{package}
\name{tryout}
\alias{tryout}
\alias{tryout-package}
\title{tryout: Automatic quality control of physical, chemical and marine biological data generated by the SHARK system}
\description{
Control data sets submitted to the Swedish Oceanographic data center SHARK (<https://sharkweb.smhi.se/>). It can also be used for checking the quality of a set of marine species distribution records.
}
